Add guided authoring, viewport interaction fixes, and multi-instance MCP
This commit rolls the current Talos3D core proof-point work into one coherent platform update.

Highlights:
- add definition drafts, definition libraries, compound-definition evaluation, occurrence explain/resolve, and related MCP/model API support
- add draggable definition and material browser windows plus a Definition Inspector and guided authoring surfaces
- replace the old hidden command palette with an egui command panel wired through the public command registry
- refine viewport interaction: move now arms handle-based dragging, transform handles anchor from the grabbed marker, and snap points include authored controls plus bounds corners
- add visual snap target affordances, screenshot reliability fixes, and command / toolbar / chrome integration needed by the new workflows
- support multiple MCP-enabled Talos3D instances via per-instance ids, configurable or auto-assigned ports, discovery manifests, and get_instance_info
- document the multi-instance MCP workflow for other agents in README, AGENTS.md, and MCP model API docs

This commit is intended to leave AI agents with a clearer public surface for launching the app, discovering the right MCP endpoint, authoring reusable definitions, and manipulating geometry through the same command substrate as the UI.

+++ crates/talos3d-core/src/plugins/modeling/editable_mesh.rs
@@ -518,7 +518,7 @@ impl EditableMesh {
         // Fan triangulation (works for convex polygons)
         let mut triangles = Vec::with_capacity(verts.len() - 2);
         for i in 1..verts.len() - 1 {
-            triangles.push([verts[0], verts[i as usize] as u32, verts[i + 1] as u32]);
+            triangles.push([verts[0], verts[i], verts[i + 1]]);
         }
         triangles
     }
